app-health-back/src/@api/app-health/application-language/resolvers/app-health-get-applitacion-languages.resolver.ts
import { AppHealthGetApplitacionLanguagesHandler } from '@api/app-health/application-language';
import { AppHealthApplicationLanguage } from '@api/graphql';
import { QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Query, Resolver } from '@nestjs/graphql';

@Resolver()
export class AppHealthGetApplitacionLanguagesResolver
{
    constructor(
        private readonly handler: AppHealthGetApplitacionLanguagesHandler,
    ) {}

    @Query('appHealthGetApplitacionLanguages')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<AppHealthApplicationLanguage[]>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
        );
    }
}
